+++ EQUiSatOS/EQUiSatOS/src/sensor_drivers/sensor_def.h
@@ -161,9 +161,9 @@ typedef enum {
 	S_5VREF_ON
 } sig_id_t;
 
-uint get_low_bound_from_signal(sig_id_t sig);
-uint get_high_bound_from_signal(sig_id_t sig);
-int get_line_m_from_signal(sig_id_t sig);
-int get_line_b_from_signal(sig_id_t sig);
+uint16_t get_low_bound_from_signal(sig_id_t sig);
+uint16_t get_high_bound_from_signal(sig_id_t sig);
+uint16_t get_line_m_from_signal(sig_id_t sig); // TODO: adjust the math and function definitions to reflect this
+int16_t get_line_b_from_signal(sig_id_t sig);
 
 #endif /* SENSOR_DEF_H_ */
